// src/plugins/spotiMbed/types.ts
export type DisplayResource = Track | Album | Playlist | ArtistWithTracks | User;
// ...
export const enum RestrictionReason {
    Market = "market",
    Product = "product",
    Explicit = "explicit",
}

interface ApiResource {
    type: string;
    id: string;
    name: string;
    external_urls: { spotify: string; };
    href: string;
    uri: string;
}
export interface Pagination<T> {
    href: string;
    items: T[];
    limit: number;
    next: string | null;
    offset: number;
    previous: string | null;
    total: number;
}
export interface ResourceImage {
    height: number;
    width: number;
    url: string;
}
type Followers = { total: number; };

export enum ResourceType {
    Track = "track",
    Album = "album",
    Playlist = "playlist",
    Artist = "artist",
    User = "user",
}
export interface Track extends ApiResource {
    type: ResourceType.Track;

    album: Album<false>;
    artists: Artist[];

    restrictions?: { reason: string; };
    disc_number: number;
    duration_ms: number;
    explicit: boolean;
    is_local: boolean;
    preview_url: string | null;
    track_number: number;
    popularity: number;
    available_markets: string[];
}
export interface Album<HasTracks = true> extends ApiResource {
    type: ResourceType.Album;

    artists: Artist[];
    tracks: HasTracks extends true ? Pagination<Track> : never;

    images: ResourceImage[];
    album_type: "album" | "single" | "compilation";
    genres: string[];
    restrictions?: { reason: string; };
    release_date: string;
    release_date_precision: "year" | "month" | "day";
    label: string;
    popularity: number;
    available_markets: string[];
}
export interface PlaylistItem {
    added_at: string;
    added_by: PartialUser;
    is_local: boolean;
    track: Track | null;
}
export interface Playlist extends ApiResource {
    type: ResourceType.Playlist;

    owner: User;
    tracks: Pagination<PlaylistItem>;

    images: ResourceImage[];
    followers: Followers;
    description: string;
    collaborative: boolean;
    public: boolean;
    snapshot_id: string;
}
export interface Artist extends ApiResource {
    type: ResourceType.Artist;

    images: ResourceImage[];
    followers: Followers;
    genres: string[];
    popularity: number;
}
export interface ArtistWithTracks extends Artist {
    tracks: Track[];
}
export interface PartialUser extends ApiResource {
    type: ResourceType.User;
}
export interface User extends PartialUser {
    images: ResourceImage[];
    followers: Followers;
    display_name: string;
}